:truck: update: rename util_variables.py to constants
