+++ src/utils/__init__.py
@@ -1,4 +1,4 @@
 from .utils import is_meta_question
-from .util_variables import GROUP_MESSAGE, patterns_meta_questions
+from .constants import GROUP_MESSAGE, patterns_meta_questions
 
 __all__ = ["GROUP_MESSAGE", "is_meta_question", "patterns_meta_questions"]
